Backend/FastAPI/requirements.txt
anyio
arrow
bcrypt
certifi
cffi
charset-normalizer
click
colorama
configparser
cryptography
distlib
dnspython
ecdsa
eloquent
email-validator
fastapi
filelock
h11
httpcore
httptools
httpx
idna
inflection
itsdangerous
Jinja2
MarkupSafe
mongo
numpy
opencv-python
orjson
panda
passlib
peewee
Pillow
platformdirs
pyasn1
pycparser
pydantic
pymongo
PyMySQL
pyOpenSSL
python-dateutil
python-dotenv
python-jose
python-multipart
PyYAML
requests
rfc3986
rsa
simplejson
six
sniffio
starlette
typing_extensions
ujson
urllib3
uvicorn
virtualenv
watchfiles
websockets
xlrd
